Using assert 1.0 and updated mocha version
This updates the gem dependencies to use the latest of Assert (v1.0) and
Mocha (v0.13). There were a few deprecation changes to Mocha with requiring it.
Other then that, I chose to now include the `Assert::Mocha` module onto
`Assert::Context` instead of direction including `Mocha::API`. This is a more
robust solution and provides a hook to add additional functinality.
`Assert::Mocha` will include `Mocha::API` on the context class.